here thy come

+++ ft_strnequ.c
@@ -6,7 +6,7 @@
 /*   By: ssithole <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/06/10 10:35:24 by ssithole          #+#    #+#             */
-/*   Updated: 2019/06/10 10:41:52 by ssithole         ###   ########.fr       */
+/*   Updated: 2019/06/14 12:18:57 by ssithole         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int		ft_strnequ(char const *s1, char const *s2, size_t n)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	if (ft_strlen(s1) != ft_strlen(s2) && (!s1 || !s2))
